+++ aeryn_core/installer/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+"""One-Click Installer."""
+from .script import generate_installer_script
+__all__ = ['generate_installer_script']